add more builtins and stuff

+++ src/ast/expr.rs
@@ -6,7 +6,7 @@ use std::{fmt, rc::Rc, string::ToString};
 pub enum Type {
     Fn,
     Symbol,
-    Integer,
+    Float,
     List,
     Bool,
 }
